Fix "E395: contains argument not accepted here" when opening a YARA file (#1).
Keywords that are identical to a Vim option have to be defined via match, see [1]:

    Note that when you have a keyword that is the same as an option (even one
    that isn't allowed here), you can not use it. Use a match instead.

Although there are simpler workarounds [2], it will be better to fix this the
proper way and use `match` instead of `keyword`.

This fixes a bug accidentally introduced in commit
3559ca1b57d30ee5feedaaf6818c4a04aec93014 (the original solution worked because
`contains` was the last keyword on the line).

[1] https://vimhelp.org/syntax.txt.html#E789
[2] https://stackoverflow.com/questions/55504057/is-there-any-way-to-highlight-default-vim-keyword-without-using-match